fix: fix imports at component
