+++ src/features/profile/userManagement/index.ts
@@ -1,1 +1,10 @@
+export * from './lib/index'
+export * from './model/services/subscription.service'
+export {
+  type CostOfPaymant,
+  type CreatePaymentRequestType,
+  type CurrentPaymentType,
+  type DataResponseType,
+  type GetCostOfPaymentSubscriptionType,
+} from './model/types/services'
 export { UserManagement } from './ui/UserManagement'
